+++ src/main/java/dev/pantanal/b3/krpv/acao_social/modulos/Investment/dto/request/InvestmentCreateDto.java
@@ -11,7 +11,7 @@ public record InvestmentCreateDto(
 
         @Valid
         //@NotBlank(message = "field value_money can not be empty")
-        long value_money,
+        Double value_money,
 
         //@NotBlank(message = "field date can not be empty")
         LocalDateTime date,
